+++ CHANGES.md
@@ -26,6 +26,7 @@ Fixes:
 - `conn` and `connappct` now use ClusterSite #1733
 - fix `tenn` miss adding opinion types #1749
 - add ADDENDUM to OpinionType enum #1754
+- update `lactapp_1` parameters for 2026 #1734
 
 ## Current